When entering the chat room, the user gets a list with the connected users.
