+++ ChatService/Startup.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using ChatService.Services;
 
 namespace ChatService
 {
